refactor: refactored filestructure and where components are located
