+++ src/components/navbar/navbar.tsx
@@ -1,9 +1,8 @@
 "use client"
-import Link from "next/link"
-import React from "react"
-import { useState } from "react"
+import React, {useState} from "react"
 
 import NavbarMenu from "./navbarMenu"
+import {NavbarButton} from "@/components/navbar/navbarButton";
 
 const exampleItems: [string, string][] = [
   ["Asetukset", "/settings"],
@@ -45,21 +44,6 @@ const Navbar = () => {
         </div>
       </div>
     </div>
-  )
-}
-
-interface NavbarButtonProps {
-  link: string
-  icon: string
-}
-
-const NavbarButton = ({ link, icon }: NavbarButtonProps) => {
-  return (
-    <Link href={link} className="flex">
-      <span className="material-symbols-outlined smallPhone:h-[64px] smallPhone:text-[64px] h-[48px] w-full text-[48px] text-tokio">
-        {icon}
-      </span>
-    </Link>
   )
 }
