+++ src/components/release/releases.tsx
@@ -1,0 +1,44 @@
+import {Release} from "@/components/release/release";
+
+const Releases = () => {
+  return (
+    <div className="w-full overflow-y-auto">
+      <div className="grid grid-cols-1 items-center gap-4 sm:grid-cols-2 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+        <Release />
+      </div>
+    </div>
+  )
+}
+export default Releases
